pytest: Separate client api helpers tests
